+++ zcash_primitives/CHANGELOG.md
@@ -10,6 +10,18 @@ workspace.
 
 ## [Unreleased]
 
+### Added
+- `zcash_primitives::transaction`:
+  - `builder`:
+    - `Error::Coinbase` enum variant
+    - `impl<FE> From<coinbase::Error> for Error<FE>`
+    - `BuildConfig::Coinbase` enum variant now contains `miner_data: MinerData`
+      and `sequence: u32`.
+    - `BuildConfig::is_coinbase`
+    - `Builder::add_orchard_output`'s `value` parameter is now `Zatoshis`
+      instead of `u64`.
+    - `Builder` now supports constructing coinbase transactions.
+
 ### Changed
 - MSRV is now 1.85.1.
 
@@ -44,6 +56,7 @@ workspace.
     - `zcash_primitives::memo`
     - `zcash_primitives::zip32`
     - `zcash_primitives::legacy`
+- `zcash_primitives::transaction::builder::BuildConfig` is not `Copy` anymore.
 
 ## [0.26.1] - 2025-10-18